requirements.txt
Django~=5.0.6
djangorestframework~=3.15.1
requests~=2.31.0
python-dateutil~=2.9.0.post0